Nested rr under git-ps as a subcommand & added ls
I did this so that users could easily see their stack of commits with
`git-ps ls` in addition to requesting reviews with `git-ps rr`. Having
`git-ps` for Git patch stack also gives me the ability to add other
commands for the workflow. For example it might be nice to have
subcommands to alter the patch stack (trigger interactive rebase
properly), and to push a patch to mainline.