Exit out of request review and push if conflict
I did this so that after it cleans up after the failure to apply the
cherry pick and messages the user about what is going on it then exits
the app to prevent any other actions from being taken. This should have
been there before but I forgot it somehow.